requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
paho-mqtt==1.6.1
numpy==1.26.2
pillow==10.1.0 